Moved login/register screens to their own folder
