+++ client/src/CSS_files/Body.css
@@ -5,10 +5,6 @@
     align-items: center;
 }
 
-.BodyMenu {
-    background-color: whitesmoke;
-}
-
 .BodyTable {
     left: 0;
     right: 0;
